+++ client/views/editors-hint.js
@@ -7,64 +7,56 @@
 
 import CodeMirror from '/gen/codemirror.js'; // FIXME: generated file to make it local
 
-var HINT_ELEMENT_CLASS        = 'CodeMirror-hint';
-var ACTIVE_HINT_ELEMENT_CLASS = 'CodeMirror-hint-active';
+const HINT_ELEMENT_CLASS = 'CodeMirror-hint';
+const ACTIVE_HINT_ELEMENT_CLASS = 'CodeMirror-hint-active';
+const requestAnimationFrame = window.requestAnimationFrame || (fn => setTimeout(fn, 1000 / 60));
+const cancelAnimationFrame = window.cancelAnimationFrame || clearTimeout;
 
 CodeMirror.defineExtension('showHint', function(options) {
-    options = parseOptions(this, this.getCursor('start'), options);
-    var selections = this.listSelections();
+    options = Object.assign({
+        closeCharacters: /[\s()\[\]{};:>,]/,
+        closeOnUnfocus: true,
+        container: null
+    }, this.options.hintOptions, options);
+
+    const selections = this.listSelections();
     if (selections.length > 1) {
         return;
     }
 
     // By default, don't allow completion when something is selected.
-    // A hint function can have a `supportsSelection` property to
-    // indicate that it can handle selections.
     if (this.somethingSelected()) {
-        if (!options.hint.supportsSelection) {
-            return;
-        }
-        // Don't try with cross-line selections
-        for (var i = 0; i < selections.length; i++) {
-            if (selections[i].head.line != selections[i].anchor.line) {
-                return;
-            }
-        }
+        return;
     }
 
     if (this.state.completionActive) {
         this.state.completionActive.close();
     }
 
-    var completion = this.state.completionActive = new Completion(this, options);
-
-    if (!completion.options.hint) {
+    if (!options.hint) {
         return;
     }
+
+    const completion = this.state.completionActive = new Completion(this, options);
 
     CodeMirror.signal(this, 'startCompletion', this);
     completion.update(true);
 });
 
-function Completion(cm, options) {
-    this.cm = cm;
-    this.options = options;
-    this.widget = null;
-    this.debounce = 0;
-    this.tick = 0;
-    this.startPos = this.cm.getCursor('start');
-    this.startLen = this.cm.getLine(this.startPos.line).length - this.cm.getSelection().length;
+class Completion {
+    constructor(cm, options) {
+        this.cm = cm;
+        this.options = options;
+        this.widget = null;
+        this.debounce = 0;
+        this.tick = 0;
+        this.startPos = this.cm.getCursor('start');
+        this.startLen = this.cm.getLine(this.startPos.line).length - this.cm.getSelection().length;
 
-    cm.on('cursorActivity', this.activityFunc = () => this.cursorActivity());
-}
+        cm.on('cursorActivity', this.activityFunc = () => this.cursorActivity());
+    }
 
-var requestAnimationFrame = window.requestAnimationFrame || function(fn) {
-    return setTimeout(fn, 1000 / 60);
-};
-var cancelAnimationFrame = window.cancelAnimationFrame || clearTimeout;
-
-Completion.prototype = {
-    close: function() {
+    close() {
         if (!this.active()) {
             return;
         }
@@ -82,14 +74,15 @@ Completion.prototype = {
         }
 
         CodeMirror.signal(this.cm, 'endCompletion', this.cm);
-    },
+    }
 
-    active: function() {
-        return this.cm.state.completionActive == this;
-    },
+    active() {
+        return this.cm.state.completionActive === this;
+    }
 
-    pick: function(data, i) {
-        var completion = data.list[i];
+    pick(data, idx) {
+        const completion = data.list[idx];
+
         if (completion.hint) {
             completion.hint(this.cm, data, completion);
         } else {
@@ -103,19 +96,19 @@ Completion.prototype = {
 
         CodeMirror.signal(data, 'pick', completion);
         this.close();
-    },
+    }
 
-    cursorActivity: function() {
+    cursorActivity() {
         if (this.debounce) {
             cancelAnimationFrame(this.debounce);
             this.debounce = 0;
         }
 
-        var pos = this.cm.getCursor();
-        var line = this.cm.getLine(pos.line);
+        const pos = this.cm.getCursor();
+        const line = this.cm.getLine(pos.line);
         if (pos.line != this.startPos.line || line.length - pos.ch != this.startLen - this.startPos.ch ||
-                pos.ch < this.startPos.ch || this.cm.somethingSelected() ||
-                (!pos.ch || this.options.closeCharacters.test(line.charAt(pos.ch - 1)))) {
+            pos.ch < this.startPos.ch || this.cm.somethingSelected() ||
+            (!pos.ch || this.options.closeCharacters.test(line.charAt(pos.ch - 1)))) {
             this.close();
         } else {
             this.debounce = requestAnimationFrame(() => this.update());
@@ -123,28 +116,27 @@ Completion.prototype = {
                 this.widget.disable();
             }
         }
-    },
+    }
 
-    update: function(first) {
+    update(first) {
         if (this.tick == null) {
             return;
         }
 
-        var myTick = ++this.tick;
-
+        const myTick = ++this.tick;
         Promise.resolve(this.options.hint(this.cm, this.options)).then((data) => {
             if (this.tick == myTick) {
                 this.finishUpdate(data, first);
             }
         });
-    },
+    }
 
-    finishUpdate: function(data) {
+    finishUpdate(data) {
         if (this.data) {
             CodeMirror.signal(this.data, 'update');
         }
 
-        var picked = (this.widget && this.widget.picked);
+        const picked = this.widget && this.widget.picked;
 
         if (this.widget) {
             this.widget.close();
@@ -163,13 +155,6 @@ Completion.prototype = {
     }
 };
 
-function parseOptions(cm, pos, options) {
-    var editor = cm.options.hintOptions;
-    var out = Object.assign({}, defaultOptions, editor, options);
-
-    return out;
-}
-
 function getText(completion) {
     if (typeof completion == 'string') {
         return completion;
@@ -186,73 +171,72 @@ function getHintElement(hintsElement, el) {
     return el;
 }
 
-function Widget(completion, data) {
-    var cm = completion.cm;
-    var theme = completion.cm.options.theme;
-    var hintsEl = this.hintsEl = document.createElement('ul');
+class Widget {
+    constructor(completion, data) {
+        const cm = completion.cm;
+        const hintsEl = this.hintsEl = document.createElement('ul');
 
-    this.completion = completion;
-    this.data = data;
-    this.picked = false;
-    this.selectedHint = data.selectedHint || 0;
+        this.completion = completion;
+        this.data = data;
+        this.picked = false;
+        this.selectedHint = data.selectedHint || 0;
 
-    hintsEl.className = 'CodeMirror-hints ' + theme;
-    (completion.options.container || document.body).appendChild(hintsEl);
+        hintsEl.className = 'CodeMirror-hints ' + completion.cm.options.theme;
+        (completion.options.container || document.body).appendChild(hintsEl);
 
-    this.items = data.list.map((cur, i) => {
-        var el = hintsEl.appendChild(document.createElement('li'));
+        this.items = data.list.map((cur, idx) => {
+            const el = hintsEl.appendChild(document.createElement('li'));
 
-        el.className = HINT_ELEMENT_CLASS;
+            el.className = HINT_ELEMENT_CLASS;
 
-        if (i === this.selectedHint) {
-            el.classList.add(ACTIVE_HINT_ELEMENT_CLASS);
-        }
+            if (idx === this.selectedHint) {
+                el.classList.add(ACTIVE_HINT_ELEMENT_CLASS);
+            }
 
-        if (cur.render) {
-            cur.render(el, data, cur);
-        } else {
-            el.appendChild(document.createTextNode(cur.displayText || getText(cur)));
-        }
+            if (cur.render) {
+                cur.render(el, data, cur);
+            } else {
+                el.appendChild(document.createTextNode(cur.displayText || getText(cur)));
+            }
 
-        return el;
-    });
-
-    cm.addKeyMap(this.keyMap = {
-        Up: () => this.changeActive(this.selectedHint - 1),
-        Down: () => this.changeActive(this.selectedHint + 1),
-        Enter: () => this.pick(),
-        Tab: () => this.pick(),
-        Esc: () => completion.close()
-    });
-
-    if (completion.options.closeOnUnfocus) {
-        var closingOnBlur;
-        cm.on('blur', this.onBlur = function() {
-            closingOnBlur = setTimeout(() => completion.close(), 100);
+            return el;
         });
-        cm.on('focus', this.onFocus = () => clearTimeout(closingOnBlur));
+
+        cm.addKeyMap(this.keyMap = {
+            Up: () => this.changeActive(this.selectedHint - 1),
+            Down: () => this.changeActive(this.selectedHint + 1),
+            Enter: () => this.pick(),
+            Tab: () => this.pick(),
+            Esc: () => completion.close()
+        });
+
+        if (completion.options.closeOnUnfocus) {
+            let closingOnBlur;
+            cm.on('blur', this.onBlur = () => {
+                closingOnBlur = setTimeout(() => completion.close(), 100);
+            });
+            cm.on('focus', this.onFocus = () => clearTimeout(closingOnBlur));
+        }
+
+        this.updatePosSize();
+        document.addEventListener('scroll', this.onScroll = () => this.updatePosSize(), true);
+
+        CodeMirror.on(hintsEl, 'mousedown', () => setTimeout(() => cm.focus(), 20));
+        CodeMirror.on(hintsEl, 'click', (e) => {
+            const el = getHintElement(hintsEl, e.target);
+            const idx = this.items.indexOf(el);
+
+            if (idx !== -1) {
+                this.changeActive(idx);
+                this.pick();
+            }
+        });
+
+        CodeMirror.signal(data, 'select', data.list[this.selectedHint], this.items[this.selectedHint]);
+        return true;
     }
 
-    this.updatePosSize();
-    document.addEventListener('scroll', this.onScroll = () => this.updatePosSize(), true);
-
-    CodeMirror.on(hintsEl, 'mousedown', () => setTimeout(() => cm.focus(), 20));
-    CodeMirror.on(hintsEl, 'click', (e) => {
-        const el = getHintElement(hintsEl, e.target);
-        const idx = this.items.indexOf(el);
-
-        if (idx !== -1) {
-            this.changeActive(idx);
-            this.pick();
-        }
-    });
-
-    CodeMirror.signal(data, 'select', data.list[this.selectedHint], this.items[this.selectedHint]);
-    return true;
-}
-
-Widget.prototype = {
-    close: function() {
+    close() {
         if (this.completion.widget !== this) {
             return;
         }
@@ -261,36 +245,36 @@ Widget.prototype = {
         this.completion.cm.removeKeyMap(this.keyMap);
         this.hintsEl.remove();
 
-        var cm = this.completion.cm;
+        const cm = this.completion.cm;
         if (this.completion.options.closeOnUnfocus) {
             cm.off('blur', this.onBlur);
             cm.off('focus', this.onFocus);
         }
 
         document.removeEventListener('scroll', this.onScroll, true);
-    },
+    }
 
-    disable: function() {
+    disable() {
         this.completion.cm.removeKeyMap(this.keyMap);
         this.keyMap = {
             Enter: () => this.picked = true
         };
         this.completion.cm.addKeyMap(this.keyMap);
-    },
+    }
 
-    pick: function() {
+    pick() {
         this.completion.pick(this.data, this.selectedHint);
-    },
+    }
 
-    changeActive: function(i, avoidWrap) {
-        if (i >= this.items.length) {
-            i = avoidWrap ? this.items.length - 1 : 0;
-        } else if (i < 0) {
-            i = avoidWrap ? 0 : this.items.length - 1;
+    changeActive(idx, avoidWrap) {
+        if (idx >= this.items.length) {
+            idx = avoidWrap ? this.items.length - 1 : 0;
+        } else if (idx < 0) {
+            idx = avoidWrap ? 0 : this.items.length - 1;
         }
 
         const prev = this.items[this.selectedHint];
-        const next = this.items[this.selectedHint = i];
+        const next = this.items[this.selectedHint = idx];
 
         if (next === prev) {
             return;
@@ -309,31 +293,31 @@ Widget.prototype = {
         }
 
         CodeMirror.signal(this.data, 'select', this.data.list[this.selectedHint], next);
-    },
+    }
 
-    updatePosSize: function() {
+    updatePosSize() {
         const { completion, hintsEl, data } = this;
         const cm = completion.cm;
-        var pos = cm.cursorCoords(completion.options.alignWithWord ? data.from : null);
-        var left = pos.left;
-        var top = pos.bottom;
+        let pos = cm.cursorCoords();
+        let left = pos.left;
+        let top = pos.bottom;
         hintsEl.style.left = left + 'px';
         hintsEl.style.top = top + 'px';
         // If we're at the edge of the screen, then we want the menu to appear on the left of the cursor.
-        var winW = window.innerWidth;
-        var winH = window.innerHeight;
-        var box = hintsEl.getBoundingClientRect();
-        var overlapY = box.bottom - winH;
+        const winW = window.innerWidth;
+        const winH = window.innerHeight;
+        let box = hintsEl.getBoundingClientRect();
+        let overlapY = box.bottom - winH;
 
         if (overlapY > 0) {
-            var height = box.bottom - box.top;
-            var curTop = pos.top - (pos.bottom - box.top);
+            const height = box.bottom - box.top;
+            const curTop = pos.top - (pos.bottom - box.top);
             if (curTop - height > 0) { // Fits above cursor
                 hintsEl.style.top = (top = pos.top - height) + 'px';
             } else if (height > winH) {
                 hintsEl.style.height = (winH - 5) + 'px';
                 hintsEl.style.top = (top = pos.bottom - box.top) + 'px';
-                var cursor = cm.getCursor();
+                const cursor = cm.getCursor();
                 if (data.from.ch != cursor.ch) {
                     pos = cm.cursorCoords(cursor);
                     hintsEl.style.left = (left = pos.left) + 'px';
@@ -342,7 +326,7 @@ Widget.prototype = {
             }
         }
 
-        var overlapX = box.right - winW;
+        let overlapX = box.right - winW;
         if (overlapX > 0) {
             if (box.right - box.left > winW) {
                 hintsEl.style.width = (winW - 5) + 'px';
@@ -353,15 +337,5 @@ Widget.prototype = {
     }
 };
 
-CodeMirror.registerHelper('hint', 'auto');
 CodeMirror.commands.autocomplete = CodeMirror.showHint;
-
-var defaultOptions = {
-    hint: CodeMirror.hint.auto,
-    alignWithWord: true,
-    closeCharacters: /[\s()\[\]{};:>,]/,
-    closeOnUnfocus: true,
-    container: null
-};
-
 CodeMirror.defineOption('hintOptions', null);