+++ CHANGELOG.md
@@ -13,6 +13,7 @@
     - Removed `reportLink()` query helper (use `pageLink()` instead)
     - Added `Emitter` as a base class for `Widget`/`App`, `PageRenderer` and `ViewRenderer` classes
     - Removed `Widget#definePage()` method, use `Widget#page.define()` instead
+    - Extracted query and view editors from `report` page to a separate module, as `Widget#view.QueryEditor` and `Widget#view.ViewEditor` classes
 
 ## 1.0.0-beta.7 (23-01-2019)